requirements.txt
praw
gtts
moviepy
opencv-python
requests
selenium
ffmpeg-python
